+++ src/classes/utils/index.ts
@@ -1,0 +1,3 @@
+export { getModifiersFromProps } from './getModifiersFromProps';
+export { listModifiers } from './listModifiers';
+export { matchModifier } from './matchModifier';
