+++ src/classes/utils/matchModifier.ts
@@ -1,5 +1,5 @@
 import { isSameDay, differenceInDays } from 'date-fns';
-import { Modifier } from '../../typings';
+import { Modifier } from '../../types';
 /**
  * Return true if `day1` is after `day2`.
  */
