+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export { Caption } from './Caption';
+export { Day } from './Day';
+export { DayPicker } from './DayPicker';
+export { Head } from './Head';
+export { Month } from './Month';
+export { Navigation } from './Navigation';
+export { Week } from './Week';
+export { defaultClassNames } from './defaultClassNames';
