+++ src/components/utils/getOutsideStartDays.ts
@@ -1,6 +1,6 @@
 import { startOfWeek, differenceInDays, addDays } from 'date-fns';
-import { DateWithModifiers } from '../../classes/DateWithModifiers';
-import { DayPickerProps } from '../../typings';
+import { DateWithModifiers } from '../../classes';
+import { DayPickerProps } from '../../types';
 export function getOutsideStartDays(
   firstDay: DateWithModifiers,
   props: DayPickerProps
