+++ src/index.ts
@@ -1,14 +1,2 @@
-export { Caption } from './components/Caption';
-export { Day } from './components/Day';
-export { DayPicker } from './components/DayPicker';
-export { Head } from './components/Head';
-export { Month } from './components/Month';
-export { Navigation } from './components/Navigation';
-export { Week } from './components/Week';
-export { defaultClassNames } from './components/defaultClassNames';
-
-export {
-  prepareCaption,
-  prepareDay,
-  prepareNavigation,
-} from './components/helpers';
+export * from './components';
+export * from './components/helpers';
