--- src/types/Caption.d.ts
@@ -1,6 +1,0 @@
-import { DayPickerProps } from './DayPicker';
-
-export interface CaptionProps {
-  month: Date;
-  dayPickerProps: DayPickerProps;
-}
