--- src/types/DateWithModifers.d.ts
@@ -1,6 +1,0 @@
-import { MatchingModifiers } from './Modifiers';
-
-export interface DateWithModifiers {
-  date: Date;
-  modifiers: MatchingModifiers;
-}
