--- src/types/Day.d.ts
@@ -1,8 +1,0 @@
-import { MatchingModifiers } from './Modifiers';
-import { DayPickerProps } from './DayPicker';
-
-export interface DayProps {
-  day: Date;
-  modifiers: MatchingModifiers;
-  dayPickerProps: DayPickerProps;
-}
