--- src/types/FormatOptions.d.ts
@@ -1,3 +1,0 @@
-export interface FormatOptions {
-  locale?: Locale;
-}
