--- src/types/Head.d.ts
@@ -1,7 +1,0 @@
-import { DayPickerProps } from './DayPicker';
-
-export interface HeadProps {
-  locale: Locale;
-  showWeekNumber: boolean;
-  dayPickerProps: DayPickerProps;
-}
