--- src/types/Month.d.ts
@@ -1,6 +1,0 @@
-import { DayPickerProps } from './DayPicker';
-
-export interface MonthProps {
-  month: Date;
-  dayPickerProps: DayPickerProps;
-}
