--- src/types/Navigation.d.ts
@@ -1,5 +1,0 @@
-import { DayPickerProps } from '../typings';
-
-export interface NavigationProps {
-  dayPickerProps: DayPickerProps;
-}
