--- src/types/Week.d.ts
@@ -1,8 +1,0 @@
-import { DateWithModifiers } from '../classes/DateWithModifiers';
-import { DayPickerProps } from './DayPicker';
-
-export interface WeekProps {
-  weekNumber: number;
-  week: DateWithModifiers[];
-  dayPickerProps: DayPickerProps;
-}
