--- src/types/WeekNumber.d.ts
@@ -1,7 +1,0 @@
-import { DayPickerProps } from './DayPicker';
-
-export interface WeekNumberProps {
-  number: number;
-  days: Array<Date>;
-  dayPickerProps: DayPickerProps;
-}
