--- src/typings.d.ts
@@ -1,13 +1,0 @@
-export * from './types/Caption';
-export * from './types/ClassNames';
-export * from './types/Components';
-export * from './types/Day';
-export * from './types/DayPicker';
-export * from './types/FormatOptions';
-export * from './types/Head';
-export * from './types/Modifiers';
-export * from './types/Month';
-export * from './types/Navigation';
-export * from './types/Styles';
-export * from './types/Week';
-export * from './types/WeekNumber';
