+++ website/docs/basics.mdx
@@ -3,7 +3,7 @@ id: basics
 title: DayPicker Basics
 ---
 
-To familiarize with DayPicker, this page describes how to implement some basic date picker features.
+Familiarize with DayPicker by implementing a basic date picker for selecting days.
 
 ## Selecting a Day
